+++ Tests/RevenueCatUITests/PurchaseCompletedHandlerTests.swift
@@ -190,106 +190,7 @@ class PurchaseCompletedHandlerTests: TestCase {
         expect(error).toEventually(matchError(Self.failureError))
     }
 
-    func testHandleExternalPurchaseWithPurchaaseHandlers() throws {
-        var completed = false
-        var customPurchaseCodeExecuted = false
-
-        let purchasHandler = Self.externalPurchaseHandler { _ in
-            customPurchaseCodeExecuted = true
-            return (userCancelled: true, error: nil)
-        } performRestore: {
-            return (success: true, error: nil)
-        }
-
-        let config = PaywallViewConfiguration(purchaseHandler: purchasHandler)
-
-        try PaywallView(configuration: config).addToHierarchy()
-
-        Task {
-            _ = try await purchasHandler.purchase(package: Self.package)
-            completed = true
-        }
-
-        expect(completed).toEventually(beTrue())
-        expect(customPurchaseCodeExecuted) == true
-    }
-
-    //TODO: Make this one work?
-    func testHandleExternalRestoreWithPaywallHandlers() throws {
-        var completed = false
-        var customRestoreCodeExecuted = false
-
-        let config = PaywallViewConfiguration(customerInfo: TestData.customerInfo)
-
-        let pwv = PaywallView(configuration: config) { _ in
-            return (userCancelled: true, error: nil)
-        } performRestore: {
-            customRestoreCodeExecuted = true
-            return (success: true, error: nil)
-        }
-
-        try pwv.addToHierarchy()
-
-        Task {
-            do {
-                _ = try await pwv.purchaseHandler.restorePurchases()
-                completed = true
-            } catch {
-                print("error: \(error)")
-                print("oh nos")
-            }
-        }
-
-        expect(completed).toEventually(beTrue())
-        print(customRestoreCodeExecuted)
-        expect(customRestoreCodeExecuted) == true
-    }
-
-    func testHandleExternalRestoreWithPurchaaseHandlers() throws {
-        var completed = false
-        var customRestoreCodeExecuted = false
-
-        let purchasHandler = Self.externalPurchaseHandler { package in
-            return (userCancelled: true, error: nil)
-        } performRestore: {
-            customRestoreCodeExecuted = true
-            return (success: true, error: nil)
-        }
-
-        let config = PaywallViewConfiguration(purchaseHandler: purchasHandler)
-
-        try PaywallView(configuration: config).addToHierarchy()
-
-        Task {
-            _ = try await purchasHandler.restorePurchases()
-            completed = true
-        }
-
-        expect(completed).toEventually(beTrue())
-        expect(customRestoreCodeExecuted) == true
-    }
-
-    func testHandleExternalPurchaseWithPaywallHandlers() throws {
-        var completed = false
-        var customPurchaseCodeExecuted = false
-
-        let pwv = PaywallView() { package in
-            customPurchaseCodeExecuted = true
-            return (userCancelled: true, error: nil)
-        } performRestore: {
-            return (success: true, error: nil)
-        }
-
-        try pwv.addToHierarchy()
-
-        Task {
-            _ = try await pwv.purchaseHandler.purchase(package: Self.package)
-            completed = true
-        }
-
-        expect(completed).toEventually(beTrue())
-        expect(customPurchaseCodeExecuted) == true
-    }
+    
 
     func testOnRestoreCompleted() throws {
         var customerInfo: CustomerInfo?
@@ -335,13 +236,7 @@ class PurchaseCompletedHandlerTests: TestCase {
         expect(error).toEventually(matchError(Self.failureError))
     }
 
-    private static func externalPurchaseHandler(performPurchase: PerformPurchase? = nil,
-                                                performRestore:  PerformRestore? = nil)
-    -> PurchaseHandler {
-        .mock(purchasesAreCompletedBy: .myApp,
-              performPurchase: performPurchase,
-              performRestore: performRestore)
-    }
+
 
     private static let purchaseHandler: PurchaseHandler = .mock()
     private static let failingHandler: PurchaseHandler = .failing(failureError)
